Add UserKnownHostsFile=/dev/null to ssh options
It should avoid retry when same IP is part of known host file

```
Add correct host key in /home/prkumar/.ssh/known_hosts to get rid of this message.
Offending ED25519 key in /home/prkumar/.ssh/known_hosts:147
Password authentication is disabled to avoid man-in-the-middle attacks.
Keyboard-interactive authentication is disabled to avoid man-in-the-middle attacks.
UpdateHostkeys is disabled because the host key is not trusted.
```