added delete command for diskupload resource to fix https://github.com/crc-org/crc-cloud/issues/38
previously we were missing the command to be executed when the crcCloudImporterDiskUploadByCli resource would be deleted. As a consequence the disk image upload to S3 was not deleted from the ephemeral bucket, and the subsequent bucket deletion fail due to not empty bucket requirement.

This issue cause dangling resources after the import opertion was completed.