refactor resources inside `Apply` functions
`Apply` function is used to aggregate output values from resources, as those values are promises alike we need to wait until the resources are created to get the values and aggregate them for being used as input for another resources. The resources created inside an `Apply` function  can not be evaluated during the preview phase and it seems this may impact on execution phase (running an update when it is not needed).

This PR will remove any resources from inside the Apply functions and use them only for generating new outputs (as aggreagate ouput values)