Fixes #164. The error was hitting on aws create operation when no `aws-disk-size` parameter was set.
This behaviour was hit due to the fact that all parameters were passed to the provider, no matter if they were set or not on the cmd (with empty string as param value). This commit will check if each parameter is set on the cmd before adding it to the set of params to be passed to the aws provider.