Add parameter to set global pull-secret secret
Some projects might want to have an access to secured registry,
but the credentials are not shared across projects.
This commit adds GLOBAL_SECURED_REGISTRIES parameter, that
will allow user to create similar pull-secret as it is done in
openshift-config namespace to be available for other projects [1].
For example, the openshift-marketplace can not start properly,
because it can not pull the image. We can add the pull-secret into the
/var/lib/kubelet/config.json, but after clustersetup.sh script finish
the Machine Config Pool would be marked as degraded (because there would
be a missmatch). Adding the pull secret globaly will help avoid issue,
that the image can not be downloaded. On the end, it can be deleted.

[1] https://docs.openshift.com/container-platform/4.15/openshift_images/managing_images/using-image-pull-secrets.html#images-allow-pods-to-reference-images-from-secure-registries_using-image-pull-secrets